move selection operations into selection.py
